requirements.txt
asgiref==3.10.0
chardet==5.2.0
charset-normalizer==3.4.4
crispy-bootstrap4==2025.6
Django==5.2.8
django-crispy-forms==2.5
django-filter==25.2
django-tables2==2.7.5
pillow==12.0.0
psycopg2==2.9.11
python-dateutil==2.9.0.post0
pytz==2025.2
reportlab==4.4.4
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
python-dotenv==1.0.1
gunicorn==23.0.0
